+++ router/streams.ts
@@ -1,5 +1,6 @@
 import { Static } from '@sinclair/typebox';
-import { BaseErrorSchemaType, Err, Result } from './result';
+import { Err, Result } from './result';
+import { BaseErrorSchemaType } from './errors';
 
 export const ReadableBrokenError: Static<BaseErrorSchemaType> = {
   code: 'READABLE_BROKEN',
